+++ cmd/xaro/main.go
@@ -1,0 +1,24 @@
+package main
+
+// "github.com/hecategames/xaro/src/system"
+
+func main() {
+	// Init a Xaro object
+	// Xaro object will store the configurations and all scenes.
+
+	// _, config := system.LoadViperConfig()
+
+	// Load the configurations using the Xaro object.
+	// opts := engo.RunOptions{
+	// 	Title:         "Xaro",
+	// 	ScaleOnResize: true,
+	// 	MSAA:          0,
+	// 	Width:         config.Window.Width,
+	// 	Height:        config.Window.Height,
+	// 	VSync:         config.Window.VSync,
+	// 	Fullscreen:    config.Window.FullScreen,
+	// }
+
+	// Run the Xaro object's Game scene.
+	// engo.Run(opts, &scene.Game{})
+}
