+++ general/general.go
@@ -1,0 +1,4 @@
+/*
+Package general will store all the common functionality that all scenes can use.
+*/
+package general
