+++ pkg/errors/errors.go
@@ -1,0 +1,4 @@
+/*
+Package errors is used for custom error types used to make debugging easier.
+*/
+package errors
