+++ pkg/logging/logging.go
@@ -1,0 +1,4 @@
+/*
+Package logging stores multiple mockable zap loggers for Xaro to use.
+*/
+package logging
