+++ pkg/player/player.go
@@ -1,0 +1,4 @@
+/*
+Package player stores the playable character that can be used in all game scenes.
+*/
+package player
