+++ pkg/scene/scene.go
@@ -1,0 +1,4 @@
+/*
+Package scene contains all the scenes used in Xaro.
+*/
+package scene
