+++ pkg/xaro/xaro.go
@@ -1,0 +1,6 @@
+/*
+Package xaro is used to store the configs and scenes for the whole game.
+
+Xaro will act as a whole game manager.
+*/
+package xaro
